requirements.txt
# Advanced Cybersecurity Sandbox Platform
# Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Security & Cryptography
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# File Type Detection
python-magic==0.4.27

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
numpy==1.26.3
pandas==2.2.0

# SHAP for model explainability
shap==0.44.0

# YAML/JSON processing
pyyaml==6.0.1
orjson==3.9.12

# Templating (Dashboard)
jinja2==3.1.3

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Environment Configuration
python-dotenv==1.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Type Checking
mypy==1.8.0

# Code Quality
black==24.1.1
ruff==0.1.14
